+++ ibrdtn/ibrdtn/ibrdtn/data/Block.h
@@ -77,7 +77,7 @@ namespace dtn
 
 			void set(ProcFlags flag, const bool &value);
 			bool get(ProcFlags flag) const;
-			const size_t& getProcessingFlags() const;
+			const uint64_t& getProcessingFlags() const;
 
 			/**
 			 * Serialize the derived block payload.
